moved installer to service package top level
