+++ main.go
@@ -19,8 +19,8 @@ import (
 	"github.com/giantswarm/draughtsman/service"
 	"github.com/giantswarm/draughtsman/service/configurer/configmap"
 	"github.com/giantswarm/draughtsman/service/deployer"
-	"github.com/giantswarm/draughtsman/service/deployer/installer/helm"
 	"github.com/giantswarm/draughtsman/service/eventer/github"
+	"github.com/giantswarm/draughtsman/service/installer/helm"
 	slacknotifier "github.com/giantswarm/draughtsman/service/notifier/slack"
 	slackspec "github.com/giantswarm/draughtsman/service/slack"
 )
