+++ service/deployer/deployer.go
@@ -8,11 +8,11 @@ import (
 	micrologger "github.com/giantswarm/microkit/logger"
 
 	"github.com/giantswarm/draughtsman/flag"
-	"github.com/giantswarm/draughtsman/service/deployer/installer"
-	installerspec "github.com/giantswarm/draughtsman/service/deployer/installer/spec"
 	"github.com/giantswarm/draughtsman/service/eventer"
 	eventerspec "github.com/giantswarm/draughtsman/service/eventer/spec"
 	httpspec "github.com/giantswarm/draughtsman/service/http"
+	"github.com/giantswarm/draughtsman/service/installer"
+	installerspec "github.com/giantswarm/draughtsman/service/installer/spec"
 	"github.com/giantswarm/draughtsman/service/notifier"
 	notifierspec "github.com/giantswarm/draughtsman/service/notifier/spec"
 	slackspec "github.com/giantswarm/draughtsman/service/slack"
